+++ client/src/game/systems/logic/door/models.ts
@@ -7,6 +7,7 @@ export interface DoorRequest {
 }
 
 export type DOOR_TOGGLE_MODE = "movement" | "vision" | "both";
+export const DOOR_TOGGLE_MODES: DOOR_TOGGLE_MODE[] = ["vision", "movement", "both"];
 
 export interface DoorOptions {
     permissions?: Permissions;