+++ CHANGELOG.md
@@ -22,6 +22,7 @@ These usually have no immediately visible impact on regular users
 -   ExtraSettings add svg not working for shapes with no prior svg properties
 -   Spawn locations loading wrong
 -   Teleport zones triggering from other floors
+-   Draw tool door permissions not saving
 
 ## [2022.1] - 2022-04-25